Client: checks for annotation types in use.
If an annotation type is in use, it cannot be updated or removed. The checks were
added for collection event, specimen link, and participant annotation types.

+++ app/assets/javascripts/admin/studies/annotationTypes/ceventAnnotTypeRemoveService.js
@@ -1,14 +1,20 @@
-define(['../../module'], function(module) {
+define(['../../module', 'underscore'], function(module, _) {
   'use strict';
 
   module.service('ceventAnnotTypeRemoveService', ceventAnnotTypeRemoveService);
 
-  ceventAnnotTypeRemoveService.$inject = ['annotationTypeRemoveService', 'CeventAnnotTypeService'];
+  ceventAnnotTypeRemoveService.$inject = [
+    'annotationTypeRemoveService',
+    'modalService',
+    'CeventAnnotTypeService'
+  ];
 
   /**
    * Removes a collection event annotation type.
    */
-  function ceventAnnotTypeRemoveService(annotationTypeRemoveService, CeventAnnotTypeService) {
+  function ceventAnnotTypeRemoveService(annotationTypeRemoveService,
+                                        modalService,
+                                        CeventAnnotTypeService) {
     var service = {
       remove: remove
     };
@@ -16,12 +22,20 @@ define(['../../module'], function(module) {
 
     //-------
 
-    function remove(annotType) {
-      annotationTypeRemoveService.remove(
-        CeventAnnotTypeService.remove,
-        annotType,
-        'admin.studies.study.collection',
-        {studyId: annotType.studyId});
+    function remove(annotType, annotTypesInUse) {
+      if (_.contains(annotTypesInUse, annotType.id)) {
+        var headerText = 'Cannot remove this annotation type';
+        var bodyText = 'This annotation type is in use by a collection event type. ' +
+            'If you want to remove the annotation type, ' +
+            'it must first be removed from the collection event type(s) that use it.';
+        modalService.modalOk(headerText, bodyText);
+      } else {
+        annotationTypeRemoveService.remove(
+          CeventAnnotTypeService.remove,
+          annotType,
+          'admin.studies.study.collection',
+          {studyId: annotType.studyId});
+      }
     }
   }
