+++ app/assets/javascripts/admin/studies/annotationTypes/spcLinkAnnotTypeRemoveService.js
@@ -1,4 +1,4 @@
-define(['../../module'], function(module) {
+define(['../../module', 'underscore'], function(module, _) {
   'use strict';
 
   /**
@@ -6,9 +6,15 @@ define(['../../module'], function(module) {
    */
   module.service('spcLinkAnnotTypeRemoveService', spcLinkAnnotTypeRemoveService);
 
-  spcLinkAnnotTypeRemoveService.$inject = ['annotationTypeRemoveService', 'SpcLinkAnnotTypeService'];
+  spcLinkAnnotTypeRemoveService.$inject = [
+    'annotationTypeRemoveService',
+    'modalService',
+    'SpcLinkAnnotTypeService'
+  ];
 
-  function spcLinkAnnotTypeRemoveService(annotationTypeRemoveService, SpcLinkAnnotTypeService) {
+  function spcLinkAnnotTypeRemoveService(annotationTypeRemoveService,
+                                         modalService,
+                                         SpcLinkAnnotTypeService) {
     var service = {
       remove: remove
     };
@@ -16,12 +22,20 @@ define(['../../module'], function(module) {
 
     //-------
 
-    function remove(annotType) {
-      annotationTypeRemoveService.remove(
-        SpcLinkAnnotTypeService.remove,
-        annotType,
-        'admin.studies.study.processing',
-        {studyId: annotType.studyId});
+    function remove(annotType, annotTypesInUse) {
+      if (_.contains(annotTypesInUse, annotType.id)) {
+        var headerText = 'Cannot remove this annotation type';
+        var bodyText = 'This annotation type is in use by a specimen link type. ' +
+            'If you want to remove the annotation type, ' +
+            'it must first be removed from the speicmen link type(s) that use it.';
+        modalService.modalOk(headerText, bodyText);
+      } else {
+        annotationTypeRemoveService.remove(
+          SpcLinkAnnotTypeService.remove,
+          annotType,
+          'admin.studies.study.processing',
+          {studyId: annotType.studyId});
+      }
     }
   }
